refactor: restructure where commands are located
