+++ tosho/src/impl/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod kmkc;
 pub(crate) mod models;
 pub(crate) mod musq;
+pub(super) mod parser;
 
 /// All available implementations
 pub enum Implementations {
